+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changes
 2.5.10 (2020-09-29)
 -------------------
 - [UPD] update base addons lists
+- [ADD] Odoo 14 support
 
 2.5.9 (2020-05-25)
 ------------------